Fixing issues with moving to the new jedis
