+++ dyno-queues-redis/src/main/java/com/netflix/dyno/queues/redis/conn/DynoJedisPipe.java
@@ -21,7 +21,7 @@ package com.netflix.dyno.queues.redis.conn;
 
 import com.netflix.dyno.jedis.DynoJedisPipeline;
 import redis.clients.jedis.Response;
-import redis.clients.jedis.params.sortedset.ZAddParams;
+import redis.clients.jedis.params.ZAddParams;
 
 /**
  * @author Viren
