+++ dyno-queues-redis/src/main/java/com/netflix/dyno/queues/redis/conn/RedisPipe.java
@@ -20,7 +20,7 @@ package com.netflix.dyno.queues.redis.conn;
 
 import redis.clients.jedis.Pipeline;
 import redis.clients.jedis.Response;
-import redis.clients.jedis.params.sortedset.ZAddParams;
+import redis.clients.jedis.params.ZAddParams;
 
 /**
  * @author Viren
